--- crates/lun_codegen/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "lun_codegen"
-description = "Code generation of lun."
-version.workspace = true
-license.workspace = true
-edition = "2024"
-
-[dependencies]
-lun_ir = { workspace = true }
-lun_utils = { workspace = true }
-lun_semack = { workspace = true }
-lun_diag = { workspace = true }
