--- crates/lun_ir/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "lun_ir"
-description = "Intermidiate representation of Lun, the closest to the final assembly."
-version.workspace = true
-license.workspace = true
-edition = "2024"
-
-[dependencies]
-lun_semack = { workspace = true }
