--- crates/lun_lexer/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "lun_lexer"
-description = "Lexer for lun."
-version.workspace = true
-license.workspace = true
-edition = "2024"
-
-[dependencies]
-lun_utils = { workspace = true }
-lun_diag = { workspace = true }
-
-thiserror = { workspace = true }
