--- crates/lun_parser/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "lun_parser"
-description = "Parser for lun."
-version.workspace = true
-license.workspace = true
-edition = "2024"
-
-[dependencies]
-lun_diag = { workspace = true }
-lun_utils = { workspace = true }
-
-thiserror = { workspace = true }
