--- crates/lun_semack/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "lun_semack"
-description = "Sematic checker of lun."
-version.workspace = true
-license.workspace = true
-edition = "2024"
-
-[dependencies]
-lun_utils = { workspace = true }
-lun_diag = { workspace = true }
-lun_parser = { workspace = true }
