--- crates/lun_utils/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "lun_utils"
-description = "Utils for lun libs."
-version.workspace = true
-license.workspace = true
-edition = "2024"
-
-[dependencies]
