+++ crates/lunc_codegen/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "lunc_codegen"
+description = "Code generation of lun."
+version.workspace = true
+license.workspace = true
+edition = "2024"
+
+[dependencies]
+lunc_fir = { workspace = true }
+lunc_utils = { workspace = true }
+lunc_semack = { workspace = true }
+lunc_diag = { workspace = true }
