+++ crates/lunc_diag/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "lunc_diag"
+description = "Diagnostics for lun."
+version.workspace = true
+license.workspace = true
+edition = "2024"
+
+[dependencies]
+lunc_utils = { workspace = true }
+
+thiserror = { workspace = true }
+codespan-reporting = { workspace = true }
