+++ crates/lunc_lexer/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "lunc_lexer"
+description = "Lexer for lun."
+version.workspace = true
+license.workspace = true
+edition = "2024"
+
+[dependencies]
+lunc_utils = { workspace = true }
+lunc_diag = { workspace = true }
+
+# other dependencies
+thiserror = { workspace = true }
